Frameworks/OpenGL/mesa/bin/pick/requirements.txt
attrs==23.1.0
urwid==2.1.2
